🚚 chore: rename components and update imports
